Reorganize and pass tests
