+++ src/biokeen/convert/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+"""Conversion from BEL to proper triples."""
+
+from .converters import *
+from .io import get_triple, to_pykeen_df, to_pykeen_file
